+++ examples/ansibleplaybook-docker-execution/ansibleplaybook-docker-execution.go
@@ -20,10 +20,6 @@ func main() {
 	}
 	defer apiClient.Close()
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-=======
->>>>>>> a2a1d16 (Add example for executing Ansible playbooks in Docker)
 	executable := NewDockerExec(
 		apiClient,
 		WithEnv(
@@ -32,12 +28,6 @@ func main() {
 			},
 		),
 	)
-<<<<<<< HEAD
-=======
-	executable := NewDockerExec(apiClient)
->>>>>>> f904139 (create an example to execute Ansible within a container)
-=======
->>>>>>> a2a1d16 (Add example for executing Ansible playbooks in Docker)
 
 	ansiblePlaybookOptions := &playbook.AnsiblePlaybookOptions{
 		Connection: "local",