+++ examples/ansibleplaybook-docker-execution/dockercmd.go
@@ -11,14 +11,7 @@ import (
 	"github.com/apenella/go-ansible/v2/pkg/execute/exec"
 	"github.com/apenella/go-docker-builder/pkg/build"
 	contextpath "github.com/apenella/go-docker-builder/pkg/build/context/path"
-<<<<<<< HEAD
-<<<<<<< HEAD
-	"github.com/containerd/containerd/errdefs"
-=======
->>>>>>> f904139 (create an example to execute Ansible within a container)
-=======
-	"github.com/containerd/containerd/errdefs"
->>>>>>> a2a1d16 (Add example for executing Ansible playbooks in Docker)
+	errdefs "github.com/containerd/containerd/errdefs"
 	"github.com/docker/docker/api/types"
 	"github.com/docker/docker/api/types/container"
 	"github.com/docker/docker/api/types/mount"
@@ -27,8 +20,6 @@ import (
 	"github.com/docker/docker/pkg/stdcopy"
 )
 
-<<<<<<< HEAD
-<<<<<<< HEAD
 // dockerExecOptionsFunc is a function type that modifies dockerExec options
 type dockerExecOptionsFunc func(*DockerExec)
 
@@ -60,44 +51,6 @@ func (e *DockerExec) Options(opts ...dockerExecOptionsFunc) *DockerExec {
 		opt(e)
 	}
 	return e
-=======
-=======
-// dockerExecOptionsFunc is a function type that modifies dockerExec options
-type dockerExecOptionsFunc func(*DockerExec)
-
->>>>>>> a2a1d16 (Add example for executing Ansible playbooks in Docker)
-// DockerExec struct implements the Executable interface
-type DockerExec struct {
-	client *client.Client
-
-	Env []string
-}
-
-// Ensure DockerExec implements the Executabler interface
-var _ = execute.Executabler(&DockerExec{})
-
-// NewDockerExec creates a new DockerExec instance
-func NewDockerExec(client *client.Client, opts ...dockerExecOptionsFunc) *DockerExec {
-	exec := &DockerExec{
-		client: client,
-		Env:    []string{},
-	}
-<<<<<<< HEAD
->>>>>>> f904139 (create an example to execute Ansible within a container)
-=======
-
-	exec = exec.Options(opts...)
-
-	return exec
-}
-
-// Options applies the given options to the DockerExec instance
-func (e *DockerExec) Options(opts ...dockerExecOptionsFunc) *DockerExec {
-	for _, opt := range opts {
-		opt(e)
-	}
-	return e
->>>>>>> a2a1d16 (Add example for executing Ansible playbooks in Docker)
 }
 
 // Command is a wrapper of exec.Command
@@ -110,39 +63,20 @@ func (e *DockerExec) CommandContext(ctx context.Context, name string, arg ...str
 
 	cmd := NewDockerCmd(e.client)
 	cmd.ContainerName = "ansible_playbook_executor"
-<<<<<<< HEAD
-<<<<<<< HEAD
 
 	cmd.Env = append([]string{}, e.Env...)
-=======
-	cmd.Env = os.Environ()
-
->>>>>>> f904139 (create an example to execute Ansible within a container)
-=======
-
-	cmd.Env = append([]string{}, e.Env...)
->>>>>>> a2a1d16 (Add example for executing Ansible playbooks in Docker)
 	cmd.Cmd = append([]string{}, name)
 	cmd.Cmd = append(cmd.Cmd, arg...)
 
 	return cmd
 }
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-=======
->>>>>>> a2a1d16 (Add example for executing Ansible playbooks in Docker)
 func WithEnv(env []string) dockerExecOptionsFunc {
 	return func(exec *DockerExec) {
 		exec.Env = env
 	}
 }
 
-<<<<<<< HEAD
-=======
->>>>>>> f904139 (create an example to execute Ansible within a container)
-=======
->>>>>>> a2a1d16 (Add example for executing Ansible playbooks in Docker)
 // dockerCmdOptionsFunc is a function type that modifies dockerCmd options
 type dockerCmdOptionsFunc func(*dockerCmd)
 
@@ -155,14 +89,6 @@ type dockerCmd struct {
 	Env           []string
 	Cmd           []string
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-=======
-	AutoRemove bool
-
->>>>>>> f904139 (create an example to execute Ansible within a container)
-=======
->>>>>>> a2a1d16 (Add example for executing Ansible playbooks in Docker)
 	imagePathContext string
 	mounts           []mount.Mount
 	workingDir       string
@@ -173,14 +99,7 @@ type dockerCmd struct {
 	stderrPipeWriter io.WriteCloser
 }
 
-<<<<<<< HEAD
-<<<<<<< HEAD
 // Ensure dockerCmd implements the Cmder interface
-=======
->>>>>>> f904139 (create an example to execute Ansible within a container)
-=======
-// Ensure dockerCmd implements the Cmder interface
->>>>>>> a2a1d16 (Add example for executing Ansible playbooks in Docker)
 var _ = exec.Cmder(&dockerCmd{})
 
 // NewDockerCmd creates a new dockerCmd instance
@@ -192,13 +111,6 @@ func NewDockerCmd(client *client.Client, opts ...dockerCmdOptionsFunc) *dockerCm
 	if err != nil {
 		panic(err)
 	}
-<<<<<<< HEAD
-<<<<<<< HEAD
-=======
-	_ = ex
->>>>>>> f904139 (create an example to execute Ansible within a container)
-=======
->>>>>>> a2a1d16 (Add example for executing Ansible playbooks in Docker)
 
 	cmd := &dockerCmd{
 		client: client,
@@ -233,15 +145,7 @@ func (cmd *dockerCmd) CombinedOutput() ([]byte, error) {
 
 // Environ returns the environment variables for the command.
 func (cmd *dockerCmd) Environ() []string {
-<<<<<<< HEAD
-<<<<<<< HEAD
 	return cmd.Env
-=======
-	return nil
->>>>>>> f904139 (create an example to execute Ansible within a container)
-=======
-	return cmd.Env
->>>>>>> a2a1d16 (Add example for executing Ansible playbooks in Docker)
 }
 
 // Output runs the command and returns its standard output.
@@ -302,34 +206,16 @@ func (cmd *dockerCmd) Start() (err error) {
 		AttachStdin:  true,
 		AttachStdout: true,
 		AttachStderr: true,
-<<<<<<< HEAD
-<<<<<<< HEAD
 		Env:          cmd.Env,
-=======
->>>>>>> f904139 (create an example to execute Ansible within a container)
-=======
-		Env:          cmd.Env,
->>>>>>> a2a1d16 (Add example for executing Ansible playbooks in Docker)
 	}
 
 	containerCreateResp, err = cmd.client.ContainerCreate(
 		ctx,
 		containerConfig,
 		&container.HostConfig{
-<<<<<<< HEAD
-<<<<<<< HEAD
 			AutoRemove:     true,
 			Mounts:         cmd.mounts,
 			ReadonlyRootfs: false,
-=======
-			AutoRemove: cmd.AutoRemove,
-			Mounts:     cmd.mounts,
->>>>>>> f904139 (create an example to execute Ansible within a container)
-=======
-			AutoRemove:     true,
-			Mounts:         cmd.mounts,
-			ReadonlyRootfs: false,
->>>>>>> a2a1d16 (Add example for executing Ansible playbooks in Docker)
 		},
 		&network.NetworkingConfig{},
 		nil,
@@ -361,20 +247,10 @@ func (cmd *dockerCmd) Start() (err error) {
 		defer cmd.stdoutPipeWriter.Close()
 		defer cmd.stderrPipeWriter.Close()
 		// Copying stdout and stderr from the container to the respective pipes
-<<<<<<< HEAD
-<<<<<<< HEAD
-=======
->>>>>>> a2a1d16 (Add example for executing Ansible playbooks in Docker)
 		_, err := stdcopy.StdCopy(cmd.stdoutPipeWriter, cmd.stderrPipeWriter, attach.Reader)
 		if err != nil {
 			fmt.Println("Error copying output from container:", err)
 		}
-<<<<<<< HEAD
-=======
-		_, _ = stdcopy.StdCopy(cmd.stdoutPipeWriter, cmd.stdoutPipeWriter, attach.Reader)
->>>>>>> f904139 (create an example to execute Ansible within a container)
-=======
->>>>>>> a2a1d16 (Add example for executing Ansible playbooks in Docker)
 	}()
 
 	err = cmd.client.ContainerStart(ctx, cmd.containerID, container.StartOptions{})
@@ -419,13 +295,6 @@ func (cmd *dockerCmd) Wait() error {
 	case err = <-errCh:
 
 	case status := <-statusCh:
-<<<<<<< HEAD
-<<<<<<< HEAD
-=======
-
->>>>>>> f904139 (create an example to execute Ansible within a container)
-=======
->>>>>>> a2a1d16 (Add example for executing Ansible playbooks in Docker)
 		if status.StatusCode != 0 {
 			err = fmt.Errorf("container exited with code %d", status.StatusCode)
 		}
@@ -440,10 +309,6 @@ func (cmd *dockerCmd) Wait() error {
 }
 
 func (cmd *dockerCmd) cleanup() error {
-<<<<<<< HEAD
-<<<<<<< HEAD
-=======
->>>>>>> a2a1d16 (Add example for executing Ansible playbooks in Docker)
 	ctx := context.TODO()
 
 	// Inspect the container to check its state
@@ -470,41 +335,18 @@ func (cmd *dockerCmd) cleanup() error {
 
 	// Try to remove the container
 	err = cmd.client.ContainerRemove(ctx, cmd.containerID, container.RemoveOptions{
-<<<<<<< HEAD
-=======
-	err := cmd.client.ContainerRemove(context.TODO(), cmd.containerID, container.RemoveOptions{
->>>>>>> f904139 (create an example to execute Ansible within a container)
-=======
->>>>>>> a2a1d16 (Add example for executing Ansible playbooks in Docker)
 		Force:         true,
 		RemoveVolumes: true,
 	})
 	if err != nil {
-<<<<<<< HEAD
-<<<<<<< HEAD
-=======
->>>>>>> a2a1d16 (Add example for executing Ansible playbooks in Docker)
 		// If already removed, ignore
 		if errdefs.IsNotFound(err) {
 			fmt.Println("Container already removed")
 			return nil
 		}
-<<<<<<< HEAD
 		return fmt.Errorf("failed to remove container: %w", err)
 	}
 
 	fmt.Println("Container cleaned up successfully")
 	return nil
-=======
-=======
->>>>>>> a2a1d16 (Add example for executing Ansible playbooks in Docker)
-		return fmt.Errorf("failed to remove container: %w", err)
-	}
-
-<<<<<<< HEAD
->>>>>>> f904139 (create an example to execute Ansible within a container)
-=======
-	fmt.Println("Container cleaned up successfully")
-	return nil
->>>>>>> a2a1d16 (Add example for executing Ansible playbooks in Docker)
 }