+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unknown]
+
+### Added
+
+- New example that show how to run Ansible commands within a Docker Container [#116](https://github.com/apenella/go-ansible/issues/116)
+
 ## [2.4.1] (2025-11-13)
 
 ### Fixed