+++ RELEASE_NOTES.md
@@ -4,4 +4,4 @@
 
 ### Added
 
-- New example to run Ansible executions within a Docker Container [#116](https://github.com/apenella/go-ansible/issues/116)
+- New example that show how to run Ansible commands within a Docker Container [#116](https://github.com/apenella/go-ansible/issues/116)